memberId  long -> Integer change

+++ src/main/java/com/example/studymate/member/repository/MemberRepository.java
@@ -4,7 +4,7 @@ import com.example.studymate.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
     Optional<Member> findByEmail(String email);
     
 }
